# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Athoo Platform API - Website, CRM, Admin Panel
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List all email campaigns/logs
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Email logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLogListResponse"

    post:
      operationId: sendEmail
      tags: [emails]
      summary: Send an email campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmailInput"
      responses:
        "201":
          description: Email sent/scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailLog"
...
components:
  schemas:
...
    EmailLog:
      type: object
      required: [id, subject, status, recipientCount, createdAt]
      properties:
        id:
          type: integer
        subject:
          type: string
        body:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, scheduled, sent, failed, partial]
        recipientCount:
          type: integer
        sentCount:
          type: integer
        failedCount:
          type: integer
        templateId:
          type: ["integer", "null"]
        scheduledAt:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string
        sentByName:
          type: ["string", "null"]
        filters:
          type: ["object", "null"]
    EmailLogListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmailLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    SendEmailInput:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
        body:
          type: string
        templateId:
          type: integer
        recipientFilter:
          type: object
          properties:
            type:
              type: string
              enum: [all, providers, customers, interested, city, status, custom]
            city:
              type: string
            status:
              type: string
            leadIds:
              type: array
              items:
                type: integer
        scheduledAt:
          type: string
...