# lib/api-zod/src/generated/types/emailLog.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Athoo Platform API - Website, CRM, Admin Panel
 * OpenAPI spec version: 0.1.0
 */
import type { EmailLogFilters } from './emailLogFilters';
import type { EmailLogStatus } from './emailLogStatus';

export interface EmailLog {
  id: number;
  subject: string;
  /** @nullable */
  body?: string | null;
  status: EmailLogStatus;
  recipientCount: number;
  sentCount?: number;
  failedCount?: number;
  /** @nullable */
  templateId?: number | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  sentAt?: string | null;
  createdAt: string;
  /** @nullable */
  sentByName?: string | null;
  /** @nullable */
  filters?: EmailLogFilters;
}